feat: extract resolve script to own file
